backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
python-multipart
httpx
openai
pyjwt>=2.8.0
bcrypt>=4.0.0
python-jose[cryptography]
email-validator
